[HCF-1117] Only recover if consul data exists

+++ container-host-files/etc/hcf/config/scripts/patches/fix_consul_server_name.sh
@@ -41,9 +41,9 @@ touch "${SENTINEL}"
 # mean we may get some data loss in the event of failure of node 0 (to be
 # tested).
 if [ "${HCP_COMPONENT_INDEX}" == "0" ]; then
-  mkdir -p /var/vcap/store/consul_agent/raft/
-  touch /var/vcap/store/consul_agent/raft/peers.info
-  echo "[\"$IP_ADDRESS:8300\"]" > /var/vcap/store/consul_agent/raft/peers.json
+  if [ -f "/var/vcap/store/consul_agent/raft/peers.info" ]; then
+    echo "[\"$IP_ADDRESS:8300\"]" > /var/vcap/store/consul_agent/raft/peers.json
+  fi
 fi
 
 exit 0
